Moved O365 models+functions, hide new features, fix initial o365 trigger load bug, clean up createApplicationSetting
