+++ AzureFunctions.AngularClient/src/app/pickers/microsoft-graph/microsoft-graph.component.ts
@@ -69,8 +69,18 @@ export class MicrosoftGraphComponent {
 
         let dataRetriever = new MobileAppsClient(this._functionApp.getMainSiteUrl(), this._aiService);
         dataRetriever.retrieveOID(options, input).then(values => {
-            this._functionApp.createApplicationSetting(values.appSettingName, values.OID)
-            this.finishResourcePickup(values.appSettingName, input); // set selected drop-down item to app setting just created
+            this._globalStateService.setBusyState();
+            this._functionApp.createApplicationSetting(values.appSettingName, values.OID).subscribe(
+                r => {                   
+                    this._globalStateService.clearBusyState();
+                    this.finishResourcePickup(values.appSettingName, input); // set selected drop-down item to app setting just created
+                },
+                error => {
+                    this._globalStateService.clearBusyState();
+                    this._aiService.trackException(error, 'MicrosoftGraph component - createApplicationSetting()');
+                }
+            );
+            
         });
     }
