+++ AzureFunctions.AngularClient/src/app/shared/function-app.ts
@@ -1864,27 +1864,13 @@ export class FunctionApp {
     }
 
     // Modeled off of EventHub trigger's 'custom' tab when creating a new Event Hub connection
-    createApplicationSetting(appSettingName: string, appSettingValue: string, evtHub?: EventHubComponent) {
+    createApplicationSetting(appSettingName: string, appSettingValue: string): Observable<any> {
         if (appSettingName && appSettingValue) {
-            this._globalStateService.setBusyState();
             return this._cacheService.postArm(`${this.site.id}/config/appsettings/list`, true).flatMap(
                 r => {
-                var appSettings: ArmObj<any> = r.json();
-                appSettings.properties[appSettingName] = appSettingValue;
-                return this._cacheService.putArm(appSettings.id, this._armService.websiteApiVersion, appSettings);
-                })
-                .do(null, e => {
-                    this._globalStateService.clearBusyState();
-                    if (evtHub) {
-                        evtHub.selectInProcess = false;
-                    }
-                    this._aiService.trackException(e, 'functionApp - createApplicationSetting()');
-                })
-                .subscribe(r => {
-                    this._globalStateService.clearBusyState();
-                    if (evtHub) {
-                        evtHub.selectItem.next(appSettingName);
-                    }
+                  var appSettings: ArmObj<any> = r.json();
+                  appSettings.properties[appSettingName] = appSettingValue;
+                  return this._cacheService.putArm(appSettings.id, this._armService.websiteApiVersion, appSettings);
                 });
         }
     }
