+++ AzureFunctions.AngularClient/src/app/template-picker/template-picker.component.ts
@@ -1,4 +1,4 @@
-import {Component, Output, Input, EventEmitter, OnInit, AfterViewInit} from '@angular/core';
+﻿import {Component, Output, Input, EventEmitter, OnInit, AfterViewInit} from '@angular/core';
 import {TranslateService, TranslatePipe} from '@ngx-translate/core';
 import {Subject} from 'rxjs/Subject';
 import 'rxjs/add/operator/distinctUntilChanged';
@@ -12,7 +12,9 @@ import {GlobalStateService} from '../shared/services/global-state.service';
 import {BroadcastEvent} from '../shared/models/broadcast-event'
 import {DropDownElement} from '../shared/models/drop-down-element';
 import {PortalResources} from '../shared/models/portal-resources';
-import {Order} from '../shared/models/constants';
+import { Order } from '../shared/models/constants';
+import { MSGraphInputTypes } from '../shared/models/microsoft-graph';
+import { PortalService } from '../shared/services/portal.service';
 
 interface CategoryOrder {
     name: string;
@@ -44,6 +46,7 @@ export class TemplatePickerComponent {
     private _orderedCategoties: CategoryOrder[]= [];
     private _functionAppStream = new Subject<FunctionApp>();
     private _functionApp : FunctionApp;
+    private _MSGraphFeature: boolean = PortalService.MSGraphFeature();
 
     set template(value: string) {
         if (value) {
@@ -148,6 +151,10 @@ export class TemplatePickerComponent {
                             }
 
                             if (!this.getFilterMatach(template.metadata.filters)) {
+                                return;
+                            }
+
+                            if (template.metadata.category.includes("MSGraph") && !this._MSGraphFeature) {
                                 return;
                             }
 
@@ -291,6 +298,9 @@ export class TemplatePickerComponent {
     private getBindingTemplates(direction: DirectionType): Template[] {
         var result: Template[] = [];
         var filtered = this.bindings.filter((b) => {
+            if (!this._MSGraphFeature && b.type in MSGraphInputTypes) {
+                return false;
+            }
             return b.direction === direction;
         });
